% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forestplot.R
\name{plot_forestplot}
\alias{plot_forestplot}
\title{Plot Forest Plot from the Data.Frame Object from get_all_stats Function}
\usage{
plot_forestplot(stats)
}
\arguments{
\item{stats}{The data.frame object output from get_all_stats function. The data.frame can be further modified by the user.}
}
\description{
This function plots forest plot from the Data.Frame object from get_all_stats function
}
\examples{
get_all_stats(
  snp_name = "12:48198234:G:T",
  filelist = c(
    "CHD_TE_SUMSTAT_META_0110221.txt",
    "PVD_TE_META_X_0111221.txt"
  ),
  dataset_list = c(
    "CHD",
    "PVD"
  ),
  identifier_column_name = "MarkerName",
  effectsize_column_name = "Effect",
  stderr_column_name = "StdErr",
  pvalue_column_name = "P-value",
  samplesize_column_name = "N"
) \%>\% plot_forestplot()
}
