% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forestplot_from_file.R
\name{plot_forestplot_from_file}
\alias{plot_forestplot_from_file}
\title{Plot Forest Plot from GWAS Summary Statistics Files}
\usage{
plot_forestplot_from_file(
  snp_name = "12:48198234:G:T",
  filelist = c("study1.txt", "study2.txt"),
  dataset_list = c("STUDY1", "STUDY2"),
  identifier_column_name = "MarkerName",
  effectsize_column_name = "Effect",
  stderr_column_name = "StdErr",
  pvalue_column_name = "P-value",
  samplesize_column_name = "N",
  check_inverse_ref = FALSE
)
}
\arguments{
\item{snp_name}{The name of the SNP specified at the identifier column, usually as CHR:POS:REF:ALT}

\item{filelist}{A list of GWAS summary statistics files to process}

\item{dataset_list}{A list of names of the studies specified by the user}

\item{identifier_column_name}{The column name of the unique SNP identifier}

\item{effectsize_column_name}{The column name of estimated effect size}

\item{stderr_column_name}{The column name of estimated standard error}

\item{pvalue_column_name}{The column name of estimated P-value}

\item{samplesize_column_name}{The column name of study sample size}

\item{check_inverse_ref}{A boolean variable indicating whether to check for inversed reference and alternative allele. TRUE is only allowed for CHR:POS:REF:ALT format.}
}
\description{
This function plots forest plot from GWAS summary statistics files
}
\examples{
plot_forestplot_from_file(
  snp_name = "12:48198234:G:T",
  filelist = c(
    "CHD_TE_SUMSTAT_META_0110221.txt",
    "PVD_TE_META_X_0111221.txt"
  ),
  dataset_list = c(
    "CHD",
    "PVD"
  ),
  identifier_column_name = "MarkerName",
  effectsize_column_name = "Effect",
  stderr_column_name = "StdErr",
  pvalue_column_name = "P-value",
  samplesize_column_name = "N"
)
}
